web: fix up error display
